Added more features
This is the version that does not determine the digital separator and the thousand separator, also this version provides and output that count the frequencies of files with a certain file format AND file file structure, and also this version is not aquring the desiredobject oriented design. Further multithreading features are going to be added for writing and reading the files to be processed in/from a buffer.